requirements.txt
absl-py==2.0.0
addict==2.4.0
ansi2html==1.8.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colormaps==0.4.1
comm==0.2.0
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.0
fonttools==4.44.3
fsspec==2023.12.0
future==0.18.3
google-auth==2.24.0
google-auth-oauthlib==1.0.0
grpcio==1.59.3
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.6.0
importlib-resources==6.1.1
ipython==8.12.3
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiss-icp==0.2.10
kiwisolver==1.4.5
laspy==2.5.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
natsort==8.4.0
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==3.1
normalize==2.0.2
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.8.1.78
ouster-mapping==0.1.1
ouster-sdk==0.10.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
plyfile==1.0.2
prettytable==3.9.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
Pygments==2.17.1
pyparsing==3.1.1
pyquaternion==0.9.9
PyRECONSTRUCT==2.2.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.0
richenum==1.3.1
rpds-py==0.13.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
shellingham==1.5.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.13.0
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.225
urllib3==2.1.0
wcwidth==0.2.10
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zeroconf==0.58.2
zipp==3.17.0
